Refactor metrics & cloudwatch class
