+++ packages/lambda/src/postCloudWatchMetric/index.js
@@ -1,9 +1,8 @@
-import MetricsService from 'service/metrics'
+import { Metrics } from 'metrics'
 
 export async function handler (event, context, callback) {
-  const service = new MetricsService()
   try {
-    const metric = await service.registerMetric('CloudWatch', event.namespace, event.metricName, event.dimensions)
+    const metric = await new Metrics().registerMetric('CloudWatch', event.namespace, event.metricName, event.dimensions)
     callback(null, JSON.stringify(metric))
   } catch (error) {
     console.log(error.message)
